http/api: Move messages, sigchain, user
