+++ http/api/messages.go
@@ -1,0 +1,31 @@
+package api
+
+import "github.com/keys-pub/keys"
+
+// Message ...
+type Message struct {
+	Data []byte `json:"data"`
+	ID   string `json:"id"`
+}
+
+// MessageResponse ...
+type MessageResponse struct {
+	ID string `json:"id"`
+}
+
+// MessagesResponse is the response from messages.
+type MessagesResponse struct {
+	KID      keys.ID             `json:"kid"`
+	Messages []*Message          `json:"messages"`
+	Metadata map[string]Metadata `json:"md,omitempty"`
+	Version  string              `json:"version"`
+}
+
+// MetadataFor returns metadata for Message.
+func (r MessagesResponse) MetadataFor(msg *Message) Metadata {
+	md, ok := r.Metadata[msg.ID]
+	if !ok {
+		return Metadata{}
+	}
+	return md
+}
